+++ videos/index.php
@@ -47,13 +47,20 @@ header('Expires: 0');
     </header>
 
     <main id="main-content">
+       
         <section>
-            <h2>Video #1</h2>
-            <p>link #1</p>
+            <h2>Reports</h2>
+            <ul>
+            <li><a href="https://www.loom.com/share/930922cec4aa4664b74da23ba65fc9f7?sid=09c965aa-9918-4137-bef2-970ba6521d65">Refresh Data & Select Date Range</a></li>
+            <li>Systemwide Data</li>
+            <li>Organizations Data</li>
+            <li>Groups Data</li>
+            </ul>
         </section>
         <section>
-            <h2>Video #2</h2>
-            <p>link #2</p>
+            <h2>Settings</h2>
+            <h3>Video #1</h3>
+            <p>link #1</p>
         </section>
         <section>
             <h2>Video #3</h2>